// apps/web/src/pages/friends/Friend.types.ts
import type { Timestamp } from "firebase/firestore";

export interface IFriend {
  id: string;
  alias?: string;
  color?: string;
}

export interface IFriendList {
  [key: string]: IFriend;
}

export interface IFriendRequest {
  id: string;
  type: "sent" | "received";
  senderId: string;
  senderAlias?: string;
  senderColor?: string;
  receiverId: string;
  receiverAlias?: string;
  receiverColor?: string;
  status: "pending" | "accepted" | "rejected";
  addedAt: Timestamp;
  updatedAt: Timestamp;
}

export interface IFriendRequestRecord {
  sent: IFriendRequest[];
  received: IFriendRequest[];
}